Refactor to use err struct
