+++ src/err.rs
@@ -1,0 +1,5 @@
+#[derive(Debug)]
+pub struct CliErr {
+    pub msg: String,
+    pub code: i32,
+}
